rearranging theme folder structure
