+++ src/components/vanilla/charts/BarChart/index.tsx
@@ -4,7 +4,8 @@ import { useOverrideConfig } from '@embeddable.com/react';
 import useTimeseries from '../../../hooks/useTimeseries';
 import Container from '../../Container';
 import BarChart from './components/BarChart';
-import defaultTheme, { Theme } from '../../../../defaulttheme';
+import defaultTheme from '../../../../themes/defaulttheme';
+import { Theme } from '../../../../themes/theme';
 
 type Props = {
   clientContext?: any;
