+++ src/components/vanilla/charts/LineChart/index.tsx
@@ -27,7 +27,8 @@ import { setYAxisStepSize } from '../../../util/chartjs/common';
 import Container from '../../Container';
 import { useOverrideConfig } from '@embeddable.com/react';
 import { setChartJSDefaults } from '../../../util/chartjs/common';
-import defaultTheme, { Theme } from '../../../../defaulttheme';
+import defaultTheme from '../../../../themes/defaulttheme';
+import { Theme } from '../../../../themes/theme';
 
 ChartJS.register(
   CategoryScale,
