+++ src/components/vanilla/charts/PieChart/index.tsx
@@ -20,7 +20,8 @@ import formatValue from '../../../util/format';
 import Container from '../../Container';
 import { useOverrideConfig } from '@embeddable.com/react';
 import { setChartJSDefaults } from '../../../util/chartjs/common';
-import defaultTheme, { Theme } from '../../../../defaulttheme';
+import defaultTheme from '../../../../themes/defaulttheme';
+import { Theme } from '../../../../themes/theme';
 
 ChartJS.register(
   ChartDataLabels,
