+++ src/components/vanilla/charts/StackedAreaChart/index.tsx
@@ -27,7 +27,8 @@ import Container from '../../Container';
 import { setYAxisStepSize } from '../../../util/chartjs/common';
 import { useOverrideConfig } from '@embeddable.com/react';
 import { setChartJSDefaults } from '../../../util/chartjs/common';
-import defaultTheme, { Theme } from '../../../../defaulttheme';
+import defaultTheme from '../../../../themes/defaulttheme';
+import { Theme } from '../../../../themes/theme';
 
 ChartJS.register(
   CategoryScale,
